fix patches not applied for farmhands in highly specific cases
When loading patches through `Action: Include` which contain a Relationship token whose input arguments are set to field reference tokens like {{Target}} as a farmhand in multiplayer, then an "invalid input arguments ... $placeholder" error occurs and the patch is skipped. This only affected farmhands because the Include patch is loaded *after* the save is ready for them, at which point the token can validate the input arguments. This commit detects whether it's performing the initial load for a patch, and doesn't apply input validation yet in that case.